fix deleting old assets increment version

+++ lib/redmine_plugin_asset_pipeline/plugin_patch.rb
@@ -28,7 +28,9 @@ module RedminePluginAssetPipeline
       def mirror_assets_to_private
         source = assets_directory
         destination = private_directory
-        FileUtils.rm_rf(destination)
+        if File.exist?(destination)
+          FileUtils.rm_rf(destination)
+        end
         return unless File.directory?(source)
 
         source_files = Dir[source + "/**/*"]
